+++ tests/visibility.rs
@@ -28,8 +28,8 @@ fn all() {
 
     let client_transport = client_app.world.resource::<NetcodeClientTransport>();
     let client_id = ClientId::from_raw(client_transport.client_id());
-    let mut clients_info = server_app.world.resource_mut::<ClientsInfo>();
-    let visibility = clients_info.client_mut(client_id).visibility_mut();
+    let mut client_cache = server_app.world.resource_mut::<ClientCache>();
+    let visibility = client_cache.client_mut(client_id).visibility_mut();
     visibility.set_visibility(server_entity, false); // Shouldn't have any effect for this policy.
 
     server_app.update();
@@ -41,8 +41,8 @@ fn all() {
         .single(&client_app.world);
 
     // Reverse visibility back.
-    let mut clients_info = server_app.world.resource_mut::<ClientsInfo>();
-    let visibility = clients_info.client_mut(client_id).visibility_mut();
+    let mut client_cache = server_app.world.resource_mut::<ClientCache>();
+    let visibility = client_cache.client_mut(client_id).visibility_mut();
     visibility.set_visibility(server_entity, true);
 
     server_app.update();
@@ -105,8 +105,8 @@ fn blacklist() {
 
     let client_transport = client_app.world.resource::<NetcodeClientTransport>();
     let client_id = ClientId::from_raw(client_transport.client_id());
-    let mut clients_info = server_app.world.resource_mut::<ClientsInfo>();
-    let visibility = clients_info.client_mut(client_id).visibility_mut();
+    let mut client_cache = server_app.world.resource_mut::<ClientCache>();
+    let visibility = client_cache.client_mut(client_id).visibility_mut();
     visibility.set_visibility(server_entity, false);
 
     server_app.update();
@@ -115,8 +115,8 @@ fn blacklist() {
     assert!(client_app.world.entities().is_empty());
 
     // Reverse visibility back.
-    let mut clients_info = server_app.world.resource_mut::<ClientsInfo>();
-    let visibility = clients_info.client_mut(client_id).visibility_mut();
+    let mut client_cache = server_app.world.resource_mut::<ClientCache>();
+    let visibility = client_cache.client_mut(client_id).visibility_mut();
     visibility.set_visibility(server_entity, true);
 
     server_app.update();
@@ -150,8 +150,8 @@ fn blacklist_despawn() {
 
     let client_transport = client_app.world.resource::<NetcodeClientTransport>();
     let client_id = ClientId::from_raw(client_transport.client_id());
-    let mut clients_info = server_app.world.resource_mut::<ClientsInfo>();
-    let visibility = clients_info.client_mut(client_id).visibility_mut();
+    let mut client_cache = server_app.world.resource_mut::<ClientCache>();
+    let visibility = client_cache.client_mut(client_id).visibility_mut();
     visibility.set_visibility(server_entity, false);
     server_app.world.despawn(server_entity);
 
@@ -160,8 +160,8 @@ fn blacklist_despawn() {
 
     assert!(client_app.world.entities().is_empty());
 
-    let clients_info = server_app.world.resource::<ClientsInfo>();
-    let visibility = clients_info.client(client_id).visibility();
+    let client_cache = server_app.world.resource::<ClientCache>();
+    let visibility = client_cache.client(client_id).visibility();
     assert!(visibility.is_visible(server_entity)); // The missing entity must be removed from the list, so this should return `true`.
 }
 
@@ -216,8 +216,8 @@ fn whitelist() {
 
     let client_transport = client_app.world.resource::<NetcodeClientTransport>();
     let client_id = ClientId::from_raw(client_transport.client_id());
-    let mut clients_info = server_app.world.resource_mut::<ClientsInfo>();
-    let visibility = clients_info.client_mut(client_id).visibility_mut();
+    let mut client_cache = server_app.world.resource_mut::<ClientCache>();
+    let visibility = client_cache.client_mut(client_id).visibility_mut();
     visibility.set_visibility(server_entity, true);
 
     server_app.update();
@@ -229,8 +229,8 @@ fn whitelist() {
         .single(&client_app.world);
 
     // Reverse visibility.
-    let mut clients_info = server_app.world.resource_mut::<ClientsInfo>();
-    let visibility = clients_info.client_mut(client_id).visibility_mut();
+    let mut client_cache = server_app.world.resource_mut::<ClientCache>();
+    let visibility = client_cache.client_mut(client_id).visibility_mut();
     visibility.set_visibility(server_entity, false);
 
     server_app.update();
@@ -264,8 +264,8 @@ fn whitelist_despawn() {
 
     let client_transport = client_app.world.resource::<NetcodeClientTransport>();
     let client_id = ClientId::from_raw(client_transport.client_id());
-    let mut clients_info = server_app.world.resource_mut::<ClientsInfo>();
-    let visibility = clients_info.client_mut(client_id).visibility_mut();
+    let mut client_cache = server_app.world.resource_mut::<ClientCache>();
+    let visibility = client_cache.client_mut(client_id).visibility_mut();
     visibility.set_visibility(server_entity, true);
     server_app.world.despawn(server_entity);
 
@@ -274,8 +274,8 @@ fn whitelist_despawn() {
 
     assert!(client_app.world.entities().is_empty());
 
-    let clients_info = server_app.world.resource::<ClientsInfo>();
-    let visibility = clients_info.client(client_id).visibility();
+    let client_cache = server_app.world.resource::<ClientCache>();
+    let visibility = client_cache.client(client_id).visibility();
     assert!(!visibility.is_visible(server_entity));
 }